# Use `std::path::absolute` for hyperlinks on Windows